Observer sets are now part of the Collection module.
