--- spec/concurrent/atomic/copy_on_write_observer_set_spec.rb
@@ -1,9 +1,0 @@
-require_relative 'observer_set_shared'
-
-module Concurrent
-
-  describe CopyOnWriteObserverSet do
-    it_behaves_like 'an observer set'
-  end
-
-end
