+++ spec/concurrent/collection/copy_on_notify_observer_set_spec.rb
@@ -1,0 +1,10 @@
+require_relative 'observer_set_shared'
+
+module Concurrent
+  module Collection
+
+    describe CopyOnNotifyObserverSet do
+      it_behaves_like 'an observer set'
+    end
+  end
+end
