+++ spec/concurrent/collection/observer_set_shared.rb
